+++ NEWS.md
@@ -1,3 +1,9 @@
+# Rogue v2.2.0.9000 (development)
+
+- Improve performance of `TipInstability()`.
+- Remove `Rfast` dependency.
+
+
 # Rogue v2.2.0 (2026-03-20)
 
 ## Performance